Predefined Models of Quantum States

The submodule QuantumSavory.StatesZoo provides models for many frequently used quantum states.

Autogenerated API list for QuantumSavory.StatesZoo

QuantumSavory.StatesZoo.BarrettKokBellPair — Type

struct BarrettKokBellPair <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

ηᴬ: Individual channel transmissivity from source A to entanglement swapping station, ∈[0,1]

ηᴮ: Individual channel transmissivity from source B to entanglement swapping station, ∈[0,1]

Pᵈ: Total excess noise (photons per qubit slot) in photon detectors, ≥0, usually ≪1

ηᵈ: Detection efficiency of photon detectors, ∈[0,1]

𝒱: Mode matching parameter for individual interacting photonic pulses with |V| evaluates mode overlap and arg(V) evaluates the carrier phase mismatch, |V|∈[0,1]

m: A single parity bit determined by the click pattern (m = 0 for [0, 1, 1, 0] or [1, 0, 0, 1]; m = 1 for [1, 1, 0, 0] or [0, 0, 1, 1])

metadata

A symbolic representation of the noisy Bell pair state obtained in a Barrett-Kok style protocol (a sequence of two successful entanglement swaps), referred to as the "dual rail photonic qubit swap" in (Prajit Dhara and Guha, 2023) (see eq. C7).

See also BarrettKokBellPairW for the weighted density matrix.

QuantumSavory.StatesZoo.BarrettKokBellPair — Method

BarrettKokBellPair(ηᴬ,ηᴮ)

Asymmetric noiseless perfectly mode-matched case ηᴬ≠ηᴮ, Pᵈ=0, ηᵈ=1, 𝒱=1, m=0

QuantumSavory.StatesZoo.BarrettKokBellPair — Method

BarrettKokBellPair(η)

Symmetric noiseless perfectly mode-matched case ηᴬ=ηᴮ=η, Pᵈ=0, ηᵈ=1, 𝒱=1, m=0

QuantumSavory.StatesZoo.BarrettKokBellPairW — Type

The weighted version of BarrettKokBellPair, i.e. its trace is the probability of successfully heralding a Barrett-Kok Bell pair.

QuantumSavory.StatesZoo.DualRailMidSwapBell — Type

struct DualRailMidSwapBell <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

eA: Link efficiencies for memory A upto the swap (include link loss, detector efficiency, etc.)

eB: Link efficiencies for memory B upto the swap (include link loss, detector efficiency, etc.)

gA: Memory initialization parameter for memory A

gB: Memory initialization parameter for memory B

Pd: Detector dark count probability per photonic mode (assumed to be the same for both detectors)

Vis: Interferometer visibility for the midpoint swap

metadata

Generates the normalized spin-spin density matrix for linear photonic entanglement swap with emissive memories emitting dual rail photonic qubits from the paper (Prajit Dhara and Guha, 2023).

julia> using QuantumSavory.StatesZoo: DualRailMidSwapBell

julia> r = Register(2);

julia> initialize!(r[1:2], DualRailMidSwapBell(0.9, 0.9, 0.5, 0.5, 1e-8, 0.99));

julia> observable(r[1:2], Z⊗Z)
-0.9999999911111113 + 0.0im

QuantumSavory.StatesZoo.DualRailMidSwapBellW — Type

struct DualRailMidSwapBellW <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

eA: Link efficiencies for memory A upto the swap (include link loss, detector efficiency, etc.)

eB: Link efficiencies for memory B upto the swap (include link loss, detector efficiency, etc.)

gA: Memory initialization parameter for memory A

gB: Memory initialization parameter for memory B

Pd: Detector dark count probability per photonic mode (assumed to be the same for both detectors)

Vis: Interferometer visibility for the midpoint swap

metadata

Generates the unnormalized spin-spin density matrix for linear photonic entanglement swap with emissive memories emitting dual rail photonic qubits from the paper (Prajit Dhara and Guha, 2023). Since the matrix is 'weighted' by the probability for success, it is suffixed with a W to distinguish it from the normalized object DualRailMidSwapBell.

julia> using QuantumSavory.StatesZoo: DualRailMidSwapBellW

julia> r = Register(2);

julia> initialize!(r[1:2], DualRailMidSwapBellW(0.9, 0.9, 0.5, 0.5, 1e-8, 0.99));

julia> observable(r[1:2], Z⊗Z)
-0.10124999595000005 + 0.0im

QuantumSavory.StatesZoo.SingleRailMidSwapBell — Type

struct SingleRailMidSwapBell <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

eA: Link efficiencies for memory A upto the swap (include link loss, detector efficiency, etc.)

eB: Link efficiencies for memory B upto the swap (include link loss, detector efficiency, etc.)

gA: Memory initialization parameter for memory A

gB: Memory initialization parameter for memory B

Pd: Detector dark count probability per photonic mode (assumed to be the same for both detectors)

Vis: Interferometer visibility for the midpoint swap' can be complex to account for phase instability

metadata

Generates the normalized spin-spin density matrix for linear photonic entanglement swap with emissive memories emitting single rail photonic qubits from the paper (Prajit Dhara and Guha, 2023)

julia> using QuantumSavory.StatesZoo: SingleRailMidSwapBell

julia> r = Register(2);

julia> initialize!(r[1:2], SingleRailMidSwapBell(0.9, 0.9, 0.5, 0.5, 1e-8, 0.99));

julia> observable(r[1:2], Z⊗Z)
-0.8181818000000001 + 0.0im

QuantumSavory.StatesZoo.SingleRailMidSwapBellW — Type

struct SingleRailMidSwapBellW <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

eA: Link efficiencies for memory A upto the swap (include link loss, detector efficiency, etc.)

eB: Link efficiencies for memory B upto the swap (include link loss, detector efficiency, etc.)

gA: Memory initialization parameter for memory A

gB: Memory initialization parameter for memory B

Pd: Detector dark count probability per photonic mode (assumed to be the same for both detectors)

Vis: Interferometer visibility for the midpoint swap' can be complex to account for phase instability

metadata

Generates the unnormalized spin-spin density matrix for linear photonic entanglement swap with emissive memories emitting single rail photonic qubits from the paper (Prajit Dhara and Guha, 2023). Since the matrix is 'weighted' by the probability for success, it is suffixed with a W to distinguish it from the normalized object SingleRailMidSwapBell.

julia> using QuantumSavory.StatesZoo: SingleRailMidSwapBellW

julia> r = Register(2);

julia> initialize!(r[1:2], SingleRailMidSwapBellW(0.9, 0.9, 0.5, 0.5, 1e-8, 0.99));

julia> observable(r[1:2], Z⊗Z)
-0.202499993925 + 0.0im

QuantumSavory.StatesZoo.ZALMSpinPair — Type

struct ZALMSpinPair <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

Ns: mean photon number per mode of the cascaded source model

gA: qubit initialization parameter on Alice's side

gB: qubit initialization parameter on Bob's side

eAm: memory out-coupling efficiency for Alice's side (Allowed range: [0,1])

eBm: memory out-coupling efficiency for Bob's side (Allowed range: [0,1])

eAs: source out-coupling efficiency for Alice's side (Allowed range: [0,1])

eBs: source out-coupling efficiency for Bob's side (Allowed range: [0,1])

eD: detector efficiency (Allowed range: [0,1])

Pd: dark click probability per photonic mode on source's swap

Pdo1: dark click probability per photonic mode on Alice side swap

Pdo2: dark click probability per photonic mode on Bob side swap

VisF: product of visibilities of all three interferometers (Allowed range: [0,1])

metadata

Generate symbolic object for the normalized spin-spin density matrix for a cascaded source swapped with emissive spin memories. The cascaded source from papers (and Guha, 2022) and (and Englund, 2023) is stored in spin memories as discussed in (Prajit Dhara and Guha, 2023).

julia> using QuantumSavory.StatesZoo: ZALMSpinPair

julia> r = Register(2);

julia> initialize!(r[1:2], ZALMSpinPair(1e-3, 0.5, 0.5, 1, 1, 1, 1, 0.9, 1e-8, 1e-8, 1e-8, 0.99));

julia> observable(r[1:2], Z⊗Z)
-0.9999999911111112 + 0.0im

QuantumSavory.StatesZoo.ZALMSpinPairW — Type

struct ZALMSpinPairW <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

Ns: mean photon number per mode of the cascaded source model

gA: qubit initialization parameter on Alice's side

gB: qubit initialization parameter on Bob's side

eAm: memory out-coupling efficiency for Alice's side (Allowed range: [0,1])

eBm: memory out-coupling efficiency for Bob's side (Allowed range: [0,1])

eAs: source out-coupling efficiency for Alice's side (Allowed range: [0,1])

eBs: source out-coupling efficiency for Bob's side (Allowed range: [0,1])

eD: detector efficiency (Allowed range: [0,1])

Pd: dark click probability per photonic mode on source's swap

Pdo1: dark click probability per photonic mode on Alice side swap

Pdo2: dark click probability per photonic mode on Bob side swap

VisF: product of visibilities of all three interferometers (Allowed range: [0,1])

metadata

Generate symbolic object for the unnormalized spin-spin density matrix for a cascaded source swapped with emissive spin memories. Since the matrix is 'weighted' by the probability for success, it is suffixed with a W to distinguish it from the normalized object ZALMSpinPair. The cascaded source from papers (and Guha, 2022) and (and Englund, 2023) is stored in spin memories as discussed in (Prajit Dhara and Guha, 2023).

julia> using QuantumSavory.StatesZoo: ZALMSpinPairW

julia> r = Register(2);

julia> initialize!(r[1:2], ZALMSpinPairW(1e-3, 0.5, 0.5, 1, 1, 1, 1, 0.9, 1e-8, 1e-8, 1e-8, 0.99));

julia> observable(r[1:2], Z⊗Z)
-6.290288032211147e-9 + 0.0im

QuantumSavory.StatesZoo.stateexplorer — Function

An interactive explorer for two-qubit states. It returns a new figure.

Requires a Makie plotting backend to be imported.

QuantumSavory.StatesZoo.stateexplorer! — Function

An interactive explorer for two-qubit states. It modifies the given figure.

Requires a Makie plotting backend to be imported.

QuantumSavory.StatesZoo.stateparameters — Function

Return the "interesting" parameters that a state from the StatesZoo has. A constructor that uses only these parameters needs to exist.

Used by stateexplorer to generate the most valuable plots of figures of merit.

QuantumSavory.StatesZoo.stateparametersrange — Function

Return the valid ranges and the "good" value for all parameters listed in stateparameters.