Full API

Autogenerated API list

QuantumSavory.W — Constant

A wildcard instance for use with the tag querying functionality.

See also: query, tag!, tag_types, Wildcard

QuantumSavory.tag_types — Constant

Tag types available in the taggging and tag-querying system.

See also: query, tag!, Wildcard

QuantumSavory.❓ — Constant

A wildcard instance for use with the tag querying functionality.

See also: query, tag!, tag_types, Wildcard

QuantumSavory.AbstractBackground — Type

An abstract type for the various background processes that might be inflicted upon a Register slot, e.g. decay, dephasing, etc.

QuantumSavory.AmplitudeDamping — Type

A depolarization background.

QuantumSavory.ConstantHamiltonianEvolution — Type

Represents a Hamiltonian being applied for the given duration. See also NonInstantGate.

QuantumSavory.Depolarization — Type

A depolarization background.

QuantumSavory.NonInstantGate — Type

Represents an gate applied instantaneously followed by a waiting period. See also ConstantHamiltonianEvolution.

QuantumSavory.PauliNoise — Type

A Pauli noise background.

QuantumSavory.QuantumChannel — Type

Quantum channel for transmitting quantum states from one register to another.

Delay and background noise processes are supported.

The function put! is used to take the contents of a RegRef and put it in the channel. That state can can then be received by a register (after a delay) using the take! method.

julia> using QuantumSavory, ResumableFunctions, ConcurrentSim

julia> regA = Register(1); regB = Register(1);

julia> initialize!(regA[1], Z1);

julia> sim = Simulation();

julia> qc = QuantumChannel(sim, 10.0) # a delay of 10 units
QuantumChannel{Qubit}(Qubit(), DelayQueue{Register}(ConcurrentSim.QueueStore{Register, Int64}, 10.0), nothing)

julia> @resumable function alice_node(env, qc)
            println("Putting Alice's qubit in the channel at ", now(env))
            put!(qc, regA[1])
        end
alice_node (generic function with 1 method)

julia> @resumable function bob_node(env, qc)
            @yield take!(qc, regB[1])
            println("Taking the qubit from alice at ", now(env))
        end
bob_node (generic function with 1 method)

julia> @process alice_node(sim, qc); @process bob_node(sim, qc);

julia> run(sim)
Putting Alice's qubit in the channel at 0.0
Taking the qubit from alice at 10.0

julia> regA
Register with 1 slots: [ Qubit ]
  Slots:
    nothing

QuantumSavory.QuantumStateTrait — Type

An abstract type for the various types of states that can be given to Register slots, e.g. qubit, harmonic oscillator, etc.

QuantumSavory.Qubit — Type

Specifies that a given register slot contains qubits.

QuantumSavory.Qumode — Type

Specifies that a given register slot contains qumodes.

QuantumSavory.RegRef — Type

A reference to a Register slot, convenient for use with functions like apply!, etc.

julia> r = Register(2)
       initialize!(r[1], X₁)
       observable(r[1], X)
0.9999999999999998 + 0.0im

QuantumSavory.Register — Type

The main data structure in QuantumSavory, used to represent a quantum register in an arbitrary formalism.

QuantumSavory.RegisterNet — Type

A network of Registers with convenient graph API as well.

QuantumSavory.RegisterNet — Method

Construct a RegisterNet from a given list of Registers and a graph.

julia> graph = grid([2,2]) # from Graphs.jl
{4, 4} undirected simple Int64 graph

julia> registers = [Register(1), Register(2), Register(1), Register(2)]
4-element Vector{Register}:
 Register with 1 slots: [ Qubit ]
  Slots:
    nothing
 Register with 2 slots: [ Qubit | Qubit ]
  Slots:
    nothing
    nothing
 Register with 1 slots: [ Qubit ]
  Slots:
    nothing
 Register with 2 slots: [ Qubit | Qubit ]
  Slots:
    nothing
    nothing

julia> net = RegisterNet(graph, registers)
A network of 4 registers in a graph of 4 edges

julia> neighbors(net, 1) # from Graphs.jl
2-element Vector{Int64}:
 2
 3

QuantumSavory.RegisterNet — Method

Construct a RegisterNet from a given list of Registers, defaulting to a chain topology.

julia> net = RegisterNet([Register(2), Register(4), Register(2)])
A network of 3 registers in a graph of 2 edges

julia> neighbors(net,2) # from Graphs.jl
2-element Vector{Int64}:
 1
 3

QuantumSavory.T1Decay — Type

A background describing the T₁ decay of a two-level system.

QuantumSavory.T2Dephasing — Type

A background describing the T₂ dephasing of a two-level system.

QuantumInterface.apply! — Method

Apply a given operation on the given set of register slots.

apply!([regA, regB], [slot1, slot2], Gates.CNOT) would apply a CNOT gate on the content of the given registers at the given slots. The appropriate representation of the gate is used, depending on the formalism under which a quantum state is stored in the given registers. The Hilbert spaces of the registers are automatically joined if necessary.

QuantumInterface.traceout! — Method

Delete the given slot of the given register.

traceout!(reg, slot) would reset (perform a partial trace) over the given subsystem. The Hilbert space of the register gets automatically shrunk.

QuantumSavory.initialize! — Method

Set the state of a given set of registers.

initialize!([regA,regB], [slot1,slot2], state) would set the state of the given slots in the given registers to state. state can be any supported state representation, e.g., kets or density matrices from QuantumOptics.jl or tableaux from QuantumClifford.jl.

QuantumSavory.observable — Function

Calculate the expectation value of a quantum observable on the given register and slot.

observable([regA, regB], [slot1, slot2], obs) would calculate the expectation value of the obs observable (using the appropriate formalism, depending on the state representation in the given registers).

QuantumSavory.project_traceout! — Function

Perform a projective measurement on the given slot of the given register.

project_traceout!(reg, slot, [stateA, stateB]) performs a projective measurement, projecting on either stateA or stateB, returning the index of the subspace on which the projection happened. It assumes the list of possible states forms a basis for the Hilbert space. The Hilbert space of the register gets automatically shrunk.

A basis object can be specified on its own as well, e.g. project_traceout!(reg, slot, basis).

QuantumSavory.query — Method

A query function checking for the first slot in a register that has a given tag.

It supports wildcards (instances of Wildcard also available as the constants W or ❓ which can be entered as \:question: in the REPL).

julia> r = Register(10);
       tag!(r[1], :symbol, 2, 3);
       tag!(r[2], :symbol, 4, 5);
       tag!(r[5], Int, 4, 5);

julia> query(r, :symbol, 4, 5)
(Slot 2, SymbolIntInt(:symbol, 4, 5)::QuantumSavory.Tag)

julia> query(r, :symbol, ❓, 3)
(Slot 1, SymbolIntInt(:symbol, 2, 3)::QuantumSavory.Tag)

julia> query(r, :othersym, ❓, ❓) |> isnothing
true

julia> query(r, Float64, 4, 5) |> isnothing
true

QuantumSavory.registernetplot — Function

Draw the given registers on a given Makie axis.

Requires a Makie backend be already imported.

QuantumSavory.registernetplot_axis — Function

Draw the given registers on a given Makie axis.

Requires a Makie backend be already imported.

QuantumSavory.resourceplot_axis — Function

Draw the various resources and locks stored in the given meta-graph on a given Makie axis.

Requires a Makie backend be already imported.

QuantumSavory.subsystemcompose — Method

Ensure that the all slots of the given registers are represented by one single state object, i.e. that all the register slots are tracked in the same Hilbert space.

QuantumSavory.tag! — Method

Assign a tag to a slot in a register.

See also: query, tag_types

QuantumSavory.uptotime! — Function

Evolve all the states in a register to a given time, according to the various backgrounds that they might have.

julia> reg = Register(2, T1Decay(1.0))
Register with 2 slots: [ Qubit | Qubit ]
  Slots:
    nothing
    nothing

julia> initialize!(reg[1], X₁)
       observable(reg[1], σᶻ)
0.0 + 0.0im

julia> uptotime!(reg[1], 10)
       observable(reg[1], Z)
0.9999546000702374 + 0.0im