Full API

Autogenerated API list

QuantumSavory.AbstractBackground — Type

An abstract type for the various background processes that might be inflicted upon a Register slot, e.g. decay, dephasing, etc.

QuantumSavory.AmplitudeDamping — Type

A depolarization background.

QuantumSavory.ConstantHamiltonianEvolution — Type

Represents a Hamiltonian being applied for the given duration. See also NonInstantGate.

QuantumSavory.Depolarization — Type

A depolarization background.

QuantumSavory.NonInstantGate — Type

Represents an gate applied instantaneously followed by a waiting period. See also ConstantHamiltonianEvolution.

QuantumSavory.PauliNoise — Type

A Pauli noise background.

QuantumSavory.QuantumStateTrait — Type

An abstract type for the various types of states that can be given to Register slots, e.g. qubit, harmonic oscillator, etc.

QuantumSavory.Qubit — Type

Specifies that a given register slot contains qubits.

QuantumSavory.Qumode — Type

Specifies that a given register slot contains qumodes.

QuantumSavory.RegRef — Type

A reference to a Register slot, convenient for use with functions like apply!, etc.

julia> r = Register(2)
       initialize!(r[1], X₁)
       observable(r[1], X)
0.9999999999999998 + 0.0im

QuantumSavory.Register — Type

The main data structure in QuantumSavory, used to represent a quantum register in an arbitrary formalism.

QuantumSavory.RegisterNet — Type

A network of Registers with convenient graph API as well.

QuantumSavory.RegisterNet — Method

Construct a RegisterNet from a given list of Registers and a graph.

julia> graph = grid([2,2])
{4, 4} undirected simple Int64 graph

julia> registers = [Register(1), Register(2), Register(1), Register(2)]
4-element Vector{Register}:
 Register with 1 slots: [ Qubit ]
  Slots:
    nothing
 Register with 2 slots: [ Qubit | Qubit ]
  Slots:
    nothing
    nothing
 Register with 1 slots: [ Qubit ]
  Slots:
    nothing
 Register with 2 slots: [ Qubit | Qubit ]
  Slots:
    nothing
    nothing

julia> net = RegisterNet(graph, registers)
A network of 4 registers in a graph of 4 edges

julia> neighbors(net, 1)
2-element Vector{Int64}:
 2
 3

QuantumSavory.RegisterNet — Method

Construct a RegisterNet from a given list of Registers, defaulting to a chain topology.

julia> net = RegisterNet([Register(2), Register(4), Register(2)])
A network of 3 registers in a graph of 2 edges

julia> neighbors(net,2)
2-element Vector{Int64}:
 1
 3

QuantumSavory.T1Decay — Type

A background describing the T₁ decay of a two-level system.

QuantumSavory.T2Dephasing — Type

A background describing the T₂ dephasing of a two-level system.

QuantumInterface.apply! — Method

Apply a given operation on the given set of register slots.

apply!([regA, regB], [slot1, slot2], Gates.CNOT) would apply a CNOT gate on the content of the given registers at the given slots. The appropriate representatin of the gate is used, depending on the formalism under which a quantum state is stored in the given registers. The Hilbert spaces of the registers are automatically joined if necessary.

QuantumInterface.traceout! — Method

Delete the given slot of the given register.

traceout!(reg, slot) would reset (perform a partial trace) over the given subsystem. The Hilbert space of the register is automatically shrinked.

QuantumSavory.initialize! — Method

Set the state of a given set of registers.

initialize!([regA,regB], [slot1,slot2], state) would set the state of the given slots in the given registers to state. state can be any supported state representation, e.g., kets or density matrices from QuantumOptics.jl or tableaux from QuantumClifford.jl.

QuantumSavory.observable — Function

Calculate the expectation value of a quantum observable on the given register and slot.

observable([regA, regB], [slot1, slot2], obs) would calculate the expectation value of the obs observable (using the appropriate formalism, depending on the state representation in the given registers).

QuantumSavory.project_traceout! — Function

Perform a projective measurement on the given slot of the given register.

project_traceout!(reg, slot, [stateA, stateB]) performs a projective measurement, projecting on either stateA or stateB, returning the index of the subspace on which the projection happened. It assumes the list of possible states forms a basis for the Hilbert space. The Hilbert space of the register is automatically shrinked.

A basis object can be specified on its own as well, e.g. project_traceout!(reg, slot, basis).

QuantumSavory.registernetplot — Function

Draw the given registers on a given Makie axis.

Requires a Makie backend be already imported.

QuantumSavory.registernetplot_axis — Function

Draw the given registers on a given Makie axis.

Requires a Makie backend be already imported.

QuantumSavory.resourceplot_axis — Function

Draw the various resources and locks stored in the given meta-graph on a given Makie axis.

Requires a Makie backend be already imported.

QuantumSavory.subsystemcompose — Method

Ensure that the all slots of the given registers are represented by one single state object, i.e. that all the register slots are tracked in the same Hilbert space.

QuantumSavory.uptotime! — Function

Evolve all the states in a register to a given time, according to the various backgrounds that they might have.

julia> reg = Register(2, T1Decay(1.0))
Register with 2 slots: [ Qubit | Qubit ]
  Slots:
    nothing
    nothing

julia> initialize!(reg[1], X₁)
       observable(reg[1], σᶻ)
0.0 + 0.0im

julia> uptotime!(reg[1], 10)
       observable(reg[1], Z)
0.9999546000702374 + 0.0im