Full API

Autogenerated API list

QuantumSavory.QuantumCliffordRepresentation — Type

Representation using tableaux governed by QuantumClifford.jl

QuantumSavory.QuantumMCRepresentation — Type

Similar to QuantumOpticsRepresentation, but using trajectories instead of superoperators.

QuantumSavory.QuantumOpticsRepresentation — Type

Representation using kets, densinty matrices, and superoperators governed by QuantumOptics.jl.

QuantumSavory.Qubit — Type

Specifies that a given register slot contains qubits.

QuantumSavory.Qumode — Type

Specifies that a given register slot contains qumodes.

QuantumSavory.Register — Type

The main data structure in QuantumSavory, used to represent a quantum register in an arbitrary formalism.

QuantumSavory.T1Decay — Type

A background describing the T₁ decay of a two-level system.

QuantumSavory.T2Dephasing — Type

A background describing the T₂ dephasing of a two-level system.

QuantumSavory.apply! — Method

Apply a given operation on the given set of register slots.

apply!([regA, regB], [slot1, slot2], Gates.CNOT) would apply a CNOT gate on the content of the given registers at the given slots. The appropriate representatin of the gate is used, depending on the formalism under which a quantum state is stored in the given registers. The Hilbert spaces of the registers are automatically joined if necessary.

QuantumSavory.initialize! — Method

Set the state of a given set of registers.

initialize!([regA,regB], [slot1,slot2], state) would set the state of the given slots in the given registers to state. state can be any supported state representation, e.g., kets or density matrices from QuantumOptics.jl or tableaux from QuantumClifford.jl.

QuantumSavory.observable — Function

Calculate the expectation value of a quantum observable on the given register and slot.

observable([regA, regB], [slot1, slot2], obs) would calculate the expectation value of the obs observable (using the appropriate formalism, depending on the state representation in the given registers).

QuantumSavory.project_traceout! — Function

Perform a projective measurement on the given slot of the given register.

project_traceout!(reg, slot, [stateA, stateB]) performs a projective measurement, projecting on either stateA or stateB, returning the index of the subspace on which the projection happened. It assumes the list of possible states forms a basis for the Hilbert space. The Hilbert space of the register is automatically shrinked.

A basis object can be specified on its own as well, e.g. project_traceout!(reg, slot, basis).

QuantumSavory.registernetplot_axis — Method

Draw the given registers on a given Makie axis.

QuantumSavory.resourceplot_axis — Method

Draw the various resources and locks stored in the given meta-graph on a given Makie axis.

QuantumSavory.traceout! — Method

Delete the given slot of the given register.

traceout!(reg, slot) would reset (perform a partial trace) over the given subsystem. The Hilbert space of the register is automatically shrinked.