Available States

Autogenerated API list for QuantumSavory.StatesZoo

QuantumSavory.StatesZoo.DualRailMidSwapBell — Type

struct DualRailMidSwapBell <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

eA

eB

gA

gB

Pd

Vis

metadata

Generates the normalized spin-spin density matrix for linear photonic entanglement swap with emissive memories emitting dual rail photonic qubits from the paper (Prajit Dhara and Guha, 2023). It takes the following parameters:

eA, eB: Link efficiencies for memories A and B upto the swap (include link loss, detector efficiency, etc.)

gA, gB: Memory initialization parameter for memories A and B

Pd: Detector dark count probability per photonic mode (assumed to be the same for both detectors)

Vis: Interferometer visibility for the midpoint swap

julia> r = Register(2)

julia> initialize!(r[1:2], DualRailMidSwapBell(0.9, 0.9, 0.5, 0.5, 1e-8, 0.99))

julia> observable(r[1:2], Z⊗Z)

QuantumSavory.StatesZoo.DualRailMidSwapBellW — Type

struct DualRailMidSwapBellW <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

eA

eB

gA

gB

Pd

Vis

metadata

Generates the unnormalized spin-spin density matrix for linear photonic entanglement swap with emissive memories emitting dual rail photonic qubits from the paper (Prajit Dhara and Guha, 2023). Since the matrix is 'weighted' by the probability for success, it is suffixed with a W to distinguish it from the normalized object DualRailMidSwapBell. It takes the following parameters:

eA, eB: Link efficiencies for memories A and B upto the swap (include link loss, detector efficiency, etc.)

gA, gB: Memory initialization parameter for memories A and B

Pd: Detector dark count probability per photonic mode (assumed to be the same for both detectors)

Vis: Interferometer visibility for the midpoint swap

julia> r = Register(2)

julia> initialize!(r[1:2], DualRailMidSwapBellW(0.9, 0.9, 0.5, 0.5, 1e-8, 0.99))

julia> observable(r[1:2], Z⊗Z)

QuantumSavory.StatesZoo.SingleRailMidSwapBell — Type

struct SingleRailMidSwapBell <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

eA

eB

gA

gB

Pd

Vis

metadata

Generates the normalized spin-spin density matrix for linear photonic entanglement swap with emissive memories emitting single rail photonic qubits from the paper (Prajit Dhara and Guha, 2023) It takes the following parameters:

eA, eB: Link efficiencies for memories A and B upto the swap (include link loss, detector efficiency, etc.)

gA, gB: Memory initialization parameter for memories A and B

Pd: Detector dark count probability per photonic mode (assumed to be the same for both detectors)

Vis: Interferometer visibility for the midpoint swap' can be complex to account for phase instability

julia> r = Register(2)

julia> initialize!(r[1:2], SingleRailMidSwapBell(0.9, 0.9, 0.5, 0.5, 1e-8, 0.99))

julia> observable(r[1:2], Z⊗Z)

QuantumSavory.StatesZoo.SingleRailMidSwapBellW — Type

struct SingleRailMidSwapBellW <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

eA

eB

gA

gB

Pd

Vis

metadata

Generates the unnormalized spin-spin density matrix for linear photonic entanglement swap with emissive memories emitting single rail photonic qubits from the paper (Prajit Dhara and Guha, 2023). Since the matrix is 'weighted' by the probability for success, it is suffixed with a W to distinguish it from the normalized object SingleRailMidSwapBell. It takes the following parameters:

eA, eB: Link efficiencies for memories A and B upto the swap (include link loss, detector efficiency, etc.)

gA, gB: Memory initialization parameter for memories A and B

Pd: Detector dark count probability per photonic mode (assumed to be the same for both detectors)

Vis: Interferometer visibility for the midpoint swap' can be complex to account for phase instability

julia> r = Register(2)

julia> initialize!(r[1:2], SingleRailMidSwapBellW(0.9, 0.9, 0.5, 0.5, 1e-8, 0.99))

julia> observable(r[1:2], Z⊗Z)

QuantumSavory.StatesZoo.ZALMSpinPair — Type

struct ZALMSpinPair <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

Ns

gA

gB

eAm

eBm

eAs

eBs

eD

Pd

Pdo1

Pdo2

VisF

metadata

Generate symbolic object for the normalized spin-spin density matrix for a cascaded source swapped with emissive spin memories. The cascaded source from papers (and Guha, 2022) and (and Englund, 2023) is stored in spin memories as discussed in (Prajit Dhara and Guha, 2023). It takes the following parameters:

Ns: mean photon number per mode of the cascaded source model

gA: qubit initialization parameter on Alice's side

gB: qubit initialization parameter on Bob's side

eAm: memory out-coupling efficiency for Alice's side (Allowed range: [0,1])

eBm: memory out-coupling efficiency for Bob's side (Allowed range: [0,1])

eAs: source out-coupling efficiency for Alice's side (Allowed range: [0,1])

eBs: source out-coupling efficiency for Bob's side (Allowed range: [0,1])

eD: detector efficiency (Allowed range: [0,1])

Pd: dark click probability per photonic mode on source's swap

Pdo1: dark click probability per photonic mode on Alice side swap

Pdo2: dark click probability per photonic mode on Bob side swap

VisF: product of visibilities of all three interferometers (Allowed range: [0,1])

julia> r = Register(2)

julia> initialize!(r[1:2], ZALMSpinPair(1e-3, 0.5, 0.5, 1, 1, 1, 1, 0.9, 1e-8, 1e-8, 1e-8, 0.99))

juilia> observable(r[1:2], Z⊗Z)

QuantumSavory.StatesZoo.ZALMSpinPairW — Type

struct ZALMSpinPairW <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

Ns

gA

gB

eAm

eBm

eAs

eBs

eD

Pd

Pdo1

Pdo2

VisF

metadata

Generate symbolic object for the unnormalized spin-spin density matrix for a cascaded source swapped with emissive spin memories. Since the matrix is 'weighted' by the probability for success, it is suffixed with a W to distinguish it from the normalized object ZALMSpinPair. The cascaded source from papers (and Guha, 2022) and (and Englund, 2023) is stored in spin memories as discussed in (Prajit Dhara and Guha, 2023). It takes the following parameters:

Ns: mean photon number per mode of the cascaded source model

gA: qubit initialization parameter on Alice's side

gB: qubit initialization parameter on Bob's side

eAm: memory out-coupling efficiency for Alice's side (Allowed range: [0,1])

eBm: memory out-coupling efficiency for Bob's side (Allowed range: [0,1])

eAs: source out-coupling efficiency for Alice's side (Allowed range: [0,1])

eBs: source out-coupling efficiency for Bob's side (Allowed range: [0,1])

eD: detector efficiency (Allowed range: [0,1])

Pd: dark click probability per photonic mode on source's swap

Pdo1: dark click probability per photonic mode on Alice side swap

Pdo2: dark click probability per photonic mode on Bob side swap

VisF: product of visibilities of all three interferometers (Allowed range: [0,1])

julia> r = Register(2)

julia> initialize!(r[1:2], ZALMSpinPairW(1e-3, 0.5, 0.5, 1, 1, 1, 1, 0.9, 1e-8, 1e-8, 1e-8, 0.99))

juilia> observable(r[1:2], Z⊗Z)