QuantumSymbolics.jl reference

You can also consult the complete QuantumSymbolics.jl documentation.

Autogenerated API list for QuantumSymbolics

QuantumSymbolics.CNOT έΑΦ Constant

CNOT gate

QuantumSymbolics.CPHASE έΑΦ Constant

CPHASE gate

QuantumSymbolics.Create έΑΦ Constant

Creation operator, also available as the constant a╠ΓΆεδ, in an infinite dimension Fock basis. There is no unicode dagger superscript, so we use the uparrow

QuantumSymbolics.Destroy έΑΦ Constant

Annihilation operator, also available as the constant a╠Γ, in an infinite dimension Fock basis.

QuantumSymbolics.FέΓΒ έΑΦ Constant

Single photon state

QuantumSymbolics.H έΑΦ Constant

Hadamard gate

QuantumSymbolics.I έΑΦ Constant

Identity operator in qubit basis

QuantumSymbolics.N έΑΦ Constant

Number operator, also available as the constant n╠Γ, in an infinite dimension Fock basis.

QuantumSymbolics.Pm έΑΦ Constant

Pauli "minus" operator, also available as the constant ╧ΔέΓΜ

QuantumSymbolics.Pp έΑΦ Constant

Pauli "plus" operator, also available as the constant ╧ΔέΓΛ

QuantumSymbolics.X έΑΦ Constant

Pauli X operator, also available as the constant ╧Δ╦μ

QuantumSymbolics.X1 έΑΦ Constant

Basis state of ╧Δ╦μ

QuantumSymbolics.X2 έΑΦ Constant

Basis state of ╧Δ╦μ

QuantumSymbolics.Y έΑΦ Constant

Pauli Y operator, also available as the constant ╧Δ╩╕

QuantumSymbolics.Y1 έΑΦ Constant

Basis state of ╧Δ╩╕

QuantumSymbolics.Y2 έΑΦ Constant

Basis state of ╧Δ╩╕

QuantumSymbolics.Z έΑΦ Constant

Pauli Z operator, also available as the constant ╧Δά╢╗

QuantumSymbolics.Z1 έΑΦ Constant

Basis state of ╧Δά╢╗

QuantumSymbolics.Z2 έΑΦ Constant

Basis state of ╧Δά╢╗

QuantumSymbolics.vac έΑΦ Constant

Single-mode vacuum state

QuantumSymbolics.AmplifierCPTP έΑΦ Type

AmplifierCPTP(r::Number, noise::Int)

Amplification CPTP map, defined by the squeezing amplitude parameter r and thermal noise parameter noise.

QuantumSymbolics.AttenuatorCPTP έΑΦ Type

AttenuatorCPTP(theta::Number, noise::Int)

Attenuation CPTP map, defined by the beam splitter rotation parameter theta and thermal noise parameter noise.

QuantumSymbolics.BeamSplitterOp έΑΦ Type

Two-mode beamsplitter operator in defined Fock basis.

QuantumSymbolics.BosonicThermalState έΑΦ Type

Thermal bosonic state in defined Fock basis.

QuantumSymbolics.CoherentState έΑΦ Type

Coherent state in defined Fock basis.

QuantumSymbolics.CreateOp έΑΦ Type

Creation (raising) operator.

julia> f = FockState(2)
|2έθσ

julia> create = CreateOp()
aέΑι

julia> qsimplify(create*f, rewriter=qsimplify_fock)
(sqrt(3))|3έθσ

QuantumSymbolics.DephasingCPTP έΑΦ Type

Single-qubit dephasing CPTP map

QuantumSymbolics.DestroyOp έΑΦ Type

Annihilation (lowering or destroy) operator in defined Fock basis.

julia> f = FockState(2)
|2έθσ

julia> destroy = DestroyOp()
a

julia> qsimplify(destroy*f, rewriter=qsimplify_fock)
(sqrt(2))|1έθσ

QuantumSymbolics.DisplaceOp έΑΦ Type

Displacement operator in defined Fock basis.

julia> f = FockState(0)
|0έθσ

julia> displace = DisplaceOp(im)
D(im)

julia> qsimplify(displace*f, rewriter=qsimplify_fock)
|imέθσ

QuantumSymbolics.FockState έΑΦ Type

Fock state in defined Fock basis.

QuantumSymbolics.GateCPTP έΑΦ Type

A unitary gate followed by a CPTP map

QuantumSymbolics.IdentityOp έΑΦ Type

The identity operator for a given basis

julia> IdentityOp(X1έΛΩX2)
ΏζΧΑ

julia> express(IdentityOp(Z2))
Operator(dim=2x2)
  basis: Spin(1/2)sparse([1, 2], [1, 2], ComplexF64[1.0 + 0.0im, 1.0 + 0.0im], 2, 2)

QuantumSymbolics.KrausRepr έΑΦ Type

Kraus representation of a quantum channel

julia> @op AέΓΒ; @op AέΓΓ; @op AέΓΔ;

julia> K = kraus(AέΓΒ, AέΓΓ, AέΓΔ)
ΏζΤο(AέΓΒ,AέΓΓ,AέΓΔ)

julia> @op ╧Β;

julia> K*╧Β
AέΓΒ╧ΒAέΓΒέΑι+AέΓΓ╧ΒAέΓΓέΑι+AέΓΔ╧ΒAέΓΔέΑι

QuantumSymbolics.MixedState έΑΦ Type

Completely depolarized state

julia> MixedState(X1έΛΩX2)
ΏζΧΕ

julia> express(MixedState(X1έΛΩX2))
Operator(dim=4x4)
  basis: [Spin(1/2) έΛΩ Spin(1/2)]
 0.25 + 0.0im        έΜΖ             έΜΖ             έΜΖ     
       έΜΖ       0.25 + 0.0im        έΜΖ             έΜΖ
       έΜΖ             έΜΖ       0.25 + 0.0im        έΜΖ
       έΜΖ             έΜΖ             έΜΖ       0.25 + 0.0im

julia> express(MixedState(X1έΛΩX2), CliffordRepr())
ΏζΤθέΕψΏζΥΙΏζΥΚΏζΤ╢ΏζΤ╖
 
ΏζΤ│έΓΩέΦΒέΦΒ
+ X_
+ _X
ΏζΤχΏζΥΚΏζΤ╢ΏζΤ╖
 
ΏζΤ╡έΓΩέΦΒέΦΒ
+ Z_
+ _Z

QuantumSymbolics.NumberOp έΑΦ Type

Number operator.

julia> f = FockState(2)
|2έθσ

julia> num = NumberOp()
n

julia> qsimplify(num*f, rewriter=qsimplify_fock)
2|2έθσ

QuantumSymbolics.PauliNoiseCPTP έΑΦ Type

Single-qubit Pauli noise CPTP map

julia> apply!(express(Z1), [1], express(PauliNoiseCPTP(1/4,1/4,1/4)))
Operator(dim=2x2)
  basis: Spin(1/2)
 0.5+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.0im

QuantumSymbolics.PhaseShiftOp έΑΦ Type

Phase-shift operator in defined Fock basis.

julia> c = CoherentState(im)
|imέθσ

julia> phase = PhaseShiftOp(pi)
U(╧Α)

julia> qsimplify(phase*c, rewriter=qsimplify_fock)
|1.2246467991473532e-16 - 1.0imέθσ

QuantumSymbolics.QuantumToolboxRepr έΑΦ Type

Representation using kets, bras, density matrices, and superoperators governed by QuantumToolbox.jl.

QuantumSymbolics.SAdd έΑΦ Type

Addition of quantum objects (kets, operators, or bras).

julia> @ket kέΓΒ; @ket kέΓΓ;

julia> kέΓΒ + kέΓΓ
|kέΓΒέθσ+|kέΓΓέθσ

QuantumSymbolics.SAnticommutator έΑΦ Type

Symbolic anticommutator of two operators.

julia> @op A; @op B;

julia> anticommutator(A, B)
{A,B}

QuantumSymbolics.SApplyBra έΑΦ Type

Symbolic application of an operator on a bra (from the right).

julia> @bra b; @op A;

julia> b*A
έθρb|A

QuantumSymbolics.SApplyKet έΑΦ Type

Symbolic application of an operator on a ket (from the left).

julia> @ket k; @op A;

julia> A*k
A|kέθσ

QuantumSymbolics.SBra έΑΦ Type

Symbolic bra

QuantumSymbolics.SBraKet έΑΦ Type

Symbolic inner product of a bra and a ket.

julia> @bra b; @ket k;

julia> b*k
έθρb||kέθσ

QuantumSymbolics.SCommutator έΑΦ Type

Symbolic commutator of two operators.

julia> @op A; @op B;

julia> commutator(A, B)
[A,B]

julia> commutator(A, A)
ΏζθΟ

QuantumSymbolics.SConjugate έΑΦ Type

Complex conjugate of quantum objects (kets, bras, operators).

julia> @op A; @ket k;

julia> conj(A)
A╦μ

julia> conj(k)
|kέθσ╦μ

QuantumSymbolics.SDagger έΑΦ Type

Dagger, i.e., adjoint of quantum objects (kets, bras, operators).

julia> @ket a; @op A;

julia> dagger(2*im*A*a)
(0 - 2im)|aέθσέΑιAέΑι

julia> @op B;

julia> dagger(A*B)
BέΑιAέΑι

julia> έΕΜ = SHermitianOperator(:έΕΜ); U = SUnitaryOperator(:U);

julia> dagger(έΕΜ)
έΕΜ

julia> dagger(U)
UέΒ╗┬╣

QuantumSymbolics.SExpOperator έΑΦ Type

Exponential of a symbolic operator.

julia> @op A; @op B;

julia> exp(A)
exp(A)

QuantumSymbolics.SHermitianOperator έΑΦ Type

Symbolic Hermitian operator

QuantumSymbolics.SHermitianUnitaryOperator έΑΦ Type

Symbolic Hermitian and unitary operator

QuantumSymbolics.SInvOperator έΑΦ Type

Inverse of an operator.

julia> @op A;

julia> inv(A)
AέΒ╗┬╣

julia> inv(A)*A
ΏζΧΑ

QuantumSymbolics.SKet έΑΦ Type

Symbolic ket

QuantumSymbolics.SMulOperator έΑΦ Type

Symbolic application of operator on operator.

julia> @op A; @op B;

julia> A*B
AB

QuantumSymbolics.SOperator έΑΦ Type

Symbolic operator

QuantumSymbolics.SOuterKetBra έΑΦ Type

Symbolic outer product of a ket and a bra.

julia> @bra b; @ket k;

julia> k*b 
|kέθσέθρb|

QuantumSymbolics.SPartialTrace έΑΦ Type

Partial trace over system i of a composite quantum system

julia> @op ΏζΤς SpinBasis(1//2)έΛΩSpinBasis(1//2);

julia> op = ptrace(ΏζΤς, 1)
tr1(ΏζΤς)

julia> QuantumSymbolics.basis(op)
Spin(1/2)

julia> @op A; @op B;

julia> ptrace(AέΛΩB, 1)
(tr(A))B

julia> @ket k; @bra b;

julia> factorizable = A έΛΩ (k*b)
AέΛΩ|kέθσέθρb|

julia> ptrace(factorizable, 1)
(tr(A))|kέθσέθρb|

julia> ptrace(factorizable, 2)
(έθρb||kέθσ)A

julia> mixed_state = (AέΛΩ(k*b)) + ((k*b)έΛΩB)
(AέΛΩ|kέθσέθρb|)+(|kέθσέθρb|έΛΩB)

julia> ptrace(mixed_state, 1)
(tr(A))|kέθσέθρb|+(έθρb||kέθσ)B

julia> ptrace(mixed_state, 2)
(tr(B))|kέθσέθρb|+(έθρb||kέθσ)A

QuantumSymbolics.SProjector έΑΦ Type

Projector for a given ket.

julia> projector(X1έΛΩX2)
ΏζΡΠ[|XέΓΒέθσ|XέΓΓέθσ]

julia> express(projector(X2))
Operator(dim=2x2)
  basis: Spin(1/2)
  0.5+0.0im  -0.5-0.0im
 -0.5+0.0im   0.5+0.0im

QuantumSymbolics.SScaled έΑΦ Type

Scaling of a quantum object (ket, operator, or bra) by a number.

julia> @ket k
|kέθσ

julia> 2*k
2|kέθσ

julia> @op A
A

julia> 2*A
2A

QuantumSymbolics.SSuperOpApply έΑΦ Type

Symbolic application of a superoperator on an operator

julia> @op A; @superop S;

julia> S*A
S[A]

QuantumSymbolics.SSuperOperator έΑΦ Type

Symbolic superoperator

QuantumSymbolics.STrace έΑΦ Type

Trace of an operator

julia> @op A; @op B;

julia> tr(A)
tr(A)

julia> tr(commutator(A, B))
0

julia> @bra b; @ket k;

julia> tr(k*b)
έθρb||kέθσ

QuantumSymbolics.STranspose έΑΦ Type

Transpose of quantum objects (kets, bras, operators).

julia> @op A; @op B; @ket k;

julia> transpose(A)
Aά╡Α

julia> transpose(A+B)
Aά╡Α+Bά╡Α

julia> transpose(k)
|kέθσά╡Α

QuantumSymbolics.SUnitaryOperator έΑΦ Type

Symbolic unitary operator

QuantumSymbolics.SVec έΑΦ Type

Vectorization of a symbolic operator.

julia> @op A; @op B;

julia> vec(A)
|Aέθσέθσ

julia> vec(A+B)
|Aέθσέθσ+|Bέθσέθσ

QuantumSymbolics.SZeroBra έΑΦ Type

Symbolic zero bra

QuantumSymbolics.SZeroKet έΑΦ Type

Symbolic zero ket

QuantumSymbolics.SZeroOperator έΑΦ Type

Symbolic zero operator

QuantumSymbolics.SqueezeOp έΑΦ Type

Squeezing operator in defined Fock basis.

julia> S = SqueezeOp(pi)
S(╧Α)

julia> qsimplify(S*vac, rewriter=qsimplify_fock)
|0,╧Αέθσ

QuantumSymbolics.SqueezedState έΑΦ Type

Squeezed vacuum state in defined Fock basis.

QuantumSymbolics.StabilizerState έΑΦ Type

State defined by a stabilizer tableau

For full functionality you also need to import the QuantumClifford library.

julia> using QuantumClifford, QuantumOptics # needed for the internal representation of the stabilizer tableaux and the conversion to a ket

julia> StabilizerState(S"XX ZZ")
ΏζΤχέΓΓ

julia> express(StabilizerState(S"-X"))
Ket(dim=2)
  basis: Spin(1/2)
  0.7071067811865475 + 0.0im
 -0.7071067811865475 + 0.0im

QuantumSymbolics.TwoSqueezeOp έΑΦ Type

Two-mode squeezing operator in defined Fock basis.

QuantumSymbolics.TwoSqueezedState έΑΦ Type

Two-mode squeezed vacuum state, or EPR state, in defined Fock basis.

Base.conj έΑΦ Method

conj(x::Symbolic{AbstractKet})
conj(x::Symbolic{AbstractBra})
conj(x::Symbolic{AbstractOperator})
conj(x::Symbolic{AbstractSuperOperator})

Symbolic complex conjugate operation. See also SConjugate.

Base.exp έΑΦ Method

exp(x::Symbolic{AbstractOperator})

Symbolic exponential of an operator. See also SExpOperator.

Base.inv έΑΦ Method

inv(x::Symbolic{AbstractOperator})

Symbolic inverse of an operator. See also SInvOperator.

Base.transpose έΑΦ Method

transpose(x::Symbolic{AbstractKet})
transpose(x::Symbolic{AbstractBra})
transpose(x::Symbolic{AbstractOperator})

Symbolic transpose operation. See also STranspose.

Base.vec έΑΦ Method

vec(x::Symbolic{AbstractOperator})

Symbolic vector representation of an operator. See also SVec.

LinearAlgebra.tr έΑΦ Method

tr(x::Symbolic{AbstractOperator})

Symbolic trace operation. See also STrace.

QuantumInterface.dagger έΑΦ Method

dagger(x::Symbolic{AbstractBra})

Symbolic adjoint operation. See also SDagger.

QuantumInterface.dagger έΑΦ Method

dagger(x::Symbolic{AbstractKet})

Symbolic adjoint operation. See also SDagger.

QuantumInterface.dagger έΑΦ Method

dagger(x::Symbolic{AbstractOperator})

Symbolic adjoint operation. See also SDagger.

QuantumInterface.express έΑΦ Method

express(s, repr::AbstractRepresentation=QuantumOpticsRepr()[, use::AbstractUse])

The main interface for expressing symbolic quantum objects in various representations.

julia> express(X1)
Ket(dim=2)
  basis: Spin(1/2)
 0.7071067811865475 + 0.0im
 0.7071067811865475 + 0.0im

julia> express(X1, CliffordRepr())
ΏζΤθέΕψΏζΥΙΏζΥΚΏζΤ╢ΏζΤ╖
+ Z
ΏζΤχΏζΥΚΏζΤ╢ΏζΤ╖
+ X

julia> express(QuantumSymbolics.X)
Operator(dim=2x2)
  basis: Spin(1/2)
      έΜΖ       1.0 + 0.0im
 1.0 + 0.0im       έΜΖ

julia> express(QuantumSymbolics.X, CliffordRepr(), UseAsOperation())
sX

julia> express(QuantumSymbolics.X, CliffordRepr(), UseAsObservable())
+ X

QuantumInterface.projector έΑΦ Method

projector(x::Symbolic{AbstractKet})

Symbolic projection operation. See also SProjector.

QuantumInterface.ptrace έΑΦ Method

ptrace(x::Symbolic{AbstractOperator})

Symbolic partial trace operation. See also SPartialTrace.

QuantumSymbolics.anticommutator έΑΦ Function

The anticommutator of two operators.

QuantumSymbolics.commutator έΑΦ Function

The commutator of two operators.

QuantumSymbolics.consistent_representation έΑΦ Method

Pick a representation that is consistent with given representations and appropriate for the given state.

QuantumSymbolics.qexpand έΑΦ Method

qexpand(s)

Manually expand a symbolic expression of quantum objects.

julia> @op A; @op B; @op C;

julia> qexpand(commutator(A, B))
-1BA+AB

julia> qexpand(AέΛΩ(B+C))
(AέΛΩB)+(AέΛΩC)

julia> @ket kέΓΒ; @ket kέΓΓ;

julia> qexpand(A*(kέΓΒ+kέΓΓ))
A|kέΓΒέθσ+A|kέΓΓέθσ

QuantumSymbolics.qsimplify έΑΦ Method

qsimplify(s; rewriter=nothing)

Manually simplify a symbolic expression of quantum objects.

If the keyword rewriter is not specified, then qsimplify will apply every defined rule to the expression. For performance or single-purpose motivations, the user has the option to define a specific rewriter for qsimplify to apply to the expression. The defined rewriters for simplification are the following objects: - qsimplify_pauli - qsimplify_commutator - qsimplify_anticommutator - qsimplify_fock

julia> qsimplify(╧Δ╩╕*commutator(╧Δ╦μ*╧Δά╢╗, ╧Δά╢╗))
(0 - 2im)Z

julia> qsimplify(anticommutator(╧Δ╦μ, ╧Δ╦μ), rewriter=qsimplify_anticommutator)
2ΏζΧΑ

QuantumSymbolics.@bra έΑΦ Macro

@bra(name, basis=SpinBasis(1//2))

Define a symbolic bra of type SBra. By default, the defined basis is the spin-1/2 basis.

julia> @bra bέΓΒ
έθρbέΓΒ|

julia> @bra bέΓΓ FockBasis(2)
έθρbέΓΓ|

QuantumSymbolics.@ket έΑΦ Macro

@ket(name, basis=SpinBasis(1//2))

Define a symbolic ket of type SKet. By default, the defined basis is the spin-1/2 basis.

julia> @ket kέΓΒ
|kέΓΒέθσ

julia> @ket kέΓΓ FockBasis(2)
|kέΓΓέθσ

QuantumSymbolics.@op έΑΦ Macro

@op(name, basis=SpinBasis(1//2))

Define a symbolic operator of type SOperator. By default, the defined basis is the spin-1/2 basis.

julia> @op A
A

julia> @op B FockBasis(2)
B